Added buttons to sort videos based on duration and views, with option to display HD videos first
Added buttons to sort videos based on duration and views, with option to display HD videos first. This is particularly helpful when combined with the infinite scroll, so a lot of videos can be sort at once! 

I've always wanted to add a "Sort videos by popularity" function on Thisvid and here it is!